Change of product.price for resolving decimal digit debug

+++ eshop/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Product(models.Model):
     name = models.CharField(max_length=50, null=False, blank=False, unique=True)
     description = models.TextField(max_length=255, null=True, blank=True)
-    price = models.DecimalField(null=False, blank=False, decimal_places=10, max_digits=10)
+    price = models.FloatField(null=False, blank=False)
     stock = models.IntegerField(default=1, null=False, blank=False)
     created_at = models.DateTimeField(auto_now_add=True, null=False, blank=False)
